% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ellsworth_m_toohey}
\alias{ellsworth_m_toohey}
\title{Part Two: Ellsworth M Toohey}
\format{character(15)}
\usage{
ellsworth_m_toohey
}
\description{
A charater vector of each chapter in Part Two of The Fountainhead.
}
\keyword{datasets}
