% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{fountainhead}
\alias{fountainhead}
\alias{fountainhead-package}
\title{Ayn Rand's The Fountainhead}
\description{
This R package the complete text of Ayn Rand's classic 1943 novel, The Fountainhead.  The book is broken
down into 4 parts: Peter Keating, Ellsworth M. Toohey, Gail Wynard, and Howard Roark.  Each part of the book
is represented as a data set in the package.  Each data set is a character vector where each element 
represented by a chapter in the book.
}
