% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gail_wynand}
\alias{gail_wynand}
\title{Part Three: Gail Wynard}
\format{character(15)}
\usage{
gail_wynand
}
\description{
A charater vector of each chapter in Part Three of The Fountainhead.
}
\keyword{datasets}
