% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{howard_roark}
\alias{howard_roark}
\title{Part Four: Howard Roark}
\format{character(15)}
\usage{
howard_roark
}
\description{
A charater vector of each chapter in Part Four of The Fountainhead.
}
\keyword{datasets}
