% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peter_keating}
\alias{peter_keating}
\title{Part One: Peter Keating}
\format{character(15)}
\usage{
peter_keating
}
\description{
A charater vector of each chapter in Part One of The Fountainhead.
}
\keyword{datasets}
